refactor: Put all logic related to color and font settings into one component
