+++ pages/books/index.vue
@@ -1,5 +1,4 @@
 <template>
-  <Header />
   <v-container>
     <v-row>
       <v-col>
