+++ plugins/vuetify.js
@@ -1,4 +1,5 @@
 import { createVuetify } from "vuetify";
+import "vuetify/styles";
 import * as components from "vuetify/components";
 import * as directives from "vuetify/directives";
